Revert "refactor: add `markdown.ts`"
This reverts commit f0b568d74d7fbd59f123d0d52fcef81458e9f443.
