--- markdown.deps.ts
@@ -1,14 +1,0 @@
-/**
- * {@link https://github.com/remarkjs/remark/tree/remark-parse%4010.0.2/packages/remark-parse}
- */
-export { default as remarkParse } from "https://esm.sh/remark-parse@10.0.2?pin=v131&deps=unified@10.1.2";
-
-/**
- * {@link https://github.com/unifiedjs/unified}
- */
-export { unified } from "https://esm.sh/unified@10.1.2?pin=v131";
-
-/**
- * {@link https://github.com/syntax-tree/unist-util-select}
- */
-export { selectAll } from "https://esm.sh/unist-util-select@5.0.0?pin=v131";
