+++ mod.ts
@@ -1,3 +1,4 @@
+import assert from "node:assert";
 import {
   bold,
   brightYellow,
@@ -5,10 +6,37 @@ import {
   gray,
   parseStackTrace,
   red,
+  remarkParse,
+  selectAll,
+  unified,
 } from "./deps.ts";
 import { createProject, transform } from "./_transform.ts";
-import { toExtname, toMimeType } from "./_code_block.ts";
-import { parse as parseMarkdown } from "./markdown.ts";
+
+// TODO: Support tsx/jsx
+const languages = ["ts", "typescript", "js", "javascript"] as const;
+type SupportedLanguage = typeof languages[number];
+
+function toMimeType(language: SupportedLanguage): string {
+  switch (language) {
+    case "ts":
+    case "typescript":
+      return "application/typescript";
+    case "js":
+    case "javascript":
+      return "application/javascript";
+  }
+}
+
+function getExtname(language: SupportedLanguage): `.${"ts" | "js"}` {
+  switch (language) {
+    case "ts":
+    case "typescript":
+      return ".ts";
+    case "js":
+    case "javascript":
+      return ".js";
+  }
+}
 
 export type TestContext = Deno.TestContext;
 
@@ -44,22 +72,33 @@ export async function test(
   const markdown = await Deno.readTextFile(path);
   const { reporter } = options ?? {};
   const project = createProject();
-  const codeBlocks = parseMarkdown(markdown);
+  const tree = unified().use(remarkParse).parse(markdown);
+  const codeBlocks = languages.flatMap((language) =>
+    selectAll(`code[lang="${language}"]`, tree)
+  ).sort((a, b) =>
+    (a.position?.start.line ?? 0) - (b.position?.start.line ?? 0)
+  );
 
   const testSuiteName = `${kTestNamePrefix}${path}`;
   await reporter?.startTestSuite(testSuiteName);
   let status: TestResult["status"] = "success";
   for (const codeBlock of codeBlocks) {
-    const code = codeBlock.code;
-    const filename = `${crypto.randomUUID()}${toExtname(codeBlock.mediaType)}`;
+    // @ts-expect-error TODO: fix this type error.
+    const language: SupportedLanguage = codeBlock.lang;
+    const mimeType = toMimeType(language);
+
+    // @ts-expect-error TODO: fix this type error.
+    const code: string = codeBlock.value;
+    const filename = `${crypto.randomUUID()}${getExtname(language)}`;
     const transformedCode = transform({ code, filename, project });
-    const mimeType = toMimeType(codeBlock.mediaType);
     const dataURL = `data:${mimeType};base64,${
       encodeToBase64(transformedCode)
     }`;
 
-    const { range } = codeBlock;
-    const pathWithLocation = `${path}#L${range.start.line}-${range.end.line}`;
+    const position = codeBlock.position;
+    assert(position, "`position` should exist");
+    const pathWithLocation =
+      `${path}#L${position.start.line}-${position.end.line}`;
     const testCaseName = `${kTestNamePrefix}${pathWithLocation}`;
     const pass = await ctx.step(testCaseName, async () => {
       await reporter?.startTestCase(testCaseName);
